requirements.txt
blinker==1.4
cffi==1.15.0
click==8.0.3
cryptography==36.0.1
Deprecated==1.2.13
Flask==2.0.2
Flask-Cors==3.0.10
Flask-JWT-Extended==4.3.1
Flask-Mail==0.9.1
itsdangerous==2.0.1
Jinja2==3.0.3
jsonify==0.5
MarkupSafe==2.0.1
packaging==21.3
pycparser==2.21
PyJWT==2.3.0
PyMySQL==1.0.2
pyparsing==3.0.6
redis==4.1.0
six==1.16.0
Werkzeug==2.0.2
wrapt==1.13.3